[policy] check whether replacement is more incentive compatible
Avoid accepting replacements that are not more incentive compatible to
mine.  For now, as a conservative estimate, require that the minimum
between the transaction's individual feerate and ancestor feerate is
higher than the individual feerates of directly conflicting transactions
and the ancestor feerates of all original transactions.

Note that a package/transaction's ancestor feerate is not perfectly
representative of its incentive compatibility; it may overestimate (some
subset of the ancestors could be included by itself if it has other
high-feerate descendants or are themselves higher feerate than this
package/transaction). This is a conservative estimate and works for now.

Co-authored-by: Suhas Daftuar <sdaftuar@chaincode.com>