[validation] add TxValidationResult::TX_SINGLE_FAILURE, TX_UNKNOWN
With package validation rules, transactions that fail individually may
sometimes be eligible for reconsideration if submitted as part of a
package. For now, that includes trasactions that failed for being too
low fee(rate).  Add a new TxValidationResult type to distinguish these
failures from others.  In the next commits, we will abort package
validation if a tx fails for any other reason. In the future, we will
also decide how to cache failures in recent_rejects based on this
result.

Package validation also sometimes elects to skip some transactions when
it knows the package will not be submitted in order to quit sooner.