[POLICY] Add OP_2 as a standard output type
This will be used in following commits
as the flag for ephemeral anchors, allowing
dust values, but also requiring
the output to be spend in the same relay package.

We avoid the usage of OP_TRUE as that
is a common test usage, and any single
byte value push is just as good as another.