[validation] add a PreChecks loop to linearize by fee and find obvious errors
Linearization helps us understand which transactions need to be grouped
together as CPFPs. This helps us skip the low feerate parent and wait
until we see it in the high feerate child's ancestor package.
Linearization also helps us try to accept the highest feerate subset of
the package when we don't don't have room for all of it. For example, if
each of the transactions in the package share an ancestor whose
descendant limits only permit one, we should try the highest feerate
transaction(s) first.

Q: Why use PreChecks?
We need the fee and vsize of each transaction, which requires looking up
inputs. While some of the work done in PreChecks might not be necessary,
it applies fail-fast anti-DoS checks that we definitely want (e.g. a
transaction that is too big could thus cause us to load too many UTXOs).
In the future, we may also be interested in using information like the
transaction's mempool conflicts and ancestor set (calculated in
PreChecks) in the linearization or quit early logic. So PreChecks is
best for this purpose.

Q: We don't know if txns have valid signatures yet, so if something is
invalid, we might have a suboptimal linearization/grouping. Should we
regroup or retry in that case?
No. If the peer maliciously constructs a package this way, there isn't
much we can do to find the "optimal subset" without doing a ton of work.
Generally, we hope to only validate each transaction 1 time.
Instead, we just hope we have another honest peer that will send us the
"normal" package.

One of the unit tests needs to be edited so the child pays below
mempoolminfee instead of 0 fees, otherwise the whole package is aborted.