[validation] add a PreChecks loop to linearize by fee and find obvious errors
Linearization helps us understand which transactions need to be grouped
together as CPFPs. Alternative approaches:
(1) validate each transaction with its ancestor set, and allow retries
later (in another transaction's ancestor set) if they fail for being too
low fee. This results in O(n^2) validation costs if, for example, we
have a chain of 24 low-feerate transactions with a fee-bumping
descendant at the bottom.
(2) use a function other than PreChecks to grab UTXOs and calculate
fees. This results in skipping some of the anti-DoS fail-fast logic in
PreChecks, such as tx too big (which could mean loading a lot of UTXOs).
In the future, we may also be interested in using information like the
transaction's mempool conflicts and ancestor set (calculated in
PreChecks) in the linearization or quit early logic. So PreChecks is
best for this purpose.

Linearization also helps us try to accept the highest feerate subset of
the package when we don't don't have room for all of it. For example, if
each of the transactions in the package share an ancestor whose
descendant limits only permit one, we should try the highest feerate
transaction(s) first.

One of the unit tests needs to be edited so the child pays below
mempoolminfee instead of 0 fees, otherwise the whole package is aborted.