[validation] fully abort package validation for non-TX_SINGLE_FAILURE failures
When a package contains an invalid transaction, we may continue in order
to capture some valid subset of transaction(s), but we should be very
conservative about the amount of work we do.
- if a tx (or same txid) is already in mempool, skip and continue.
- if we get a TX_SINGLE_FAILURE, this tx might be CPFP'd by a
subsequent transaction in the package, so continue.
- if we get TX_CONFLICT or TX_MISSING_INPUTS, skip the affected
transaction(s) and try the others. This is to try our best to handle
a situation in which we have different chainstate from our peer(s).
- for all other failures, abort the whole thing.

Adds usage of AncestorPackage to linearize transactions and gracefully
skip ones (see cases above). This linearization is just topological for
now, but a subsequent commit will add logic to gather fee information
and use that to refine the linearization. Even though packages are
currently required to be sorted, the order may be changed, since
AncestorPackage can produce a different but valid topological sort.

Since AncestorPackage knows the size of a transaction's in-package
ancestor set, we can skip individual submission of those with in-package
dependencies, as we know they will fail for missing inputs.