[refactor] AcceptPackage loop using AncestorPackage and switch statement
Adds usage of AncestorPackage to linearize transactions and skip ones
that are submitted or already in mempool. This linearization is just
topological for now, but a subsequent commit will add logic to gather
fee information and use that to refine the linearization. Even though
packages are currently required to be sorted, the order may be changed,
since AncestorPackage can produce a different but valid topological
sort.

Use a switch statement to ensure we are handling all possible
TxValidationResults. Also simplifies the diff of later commits.