[validation] return the wtxid(s) and effective feerate if TX_SINGLE_FAILURE
With subpackage evaluation and de-duplication, it's not always the
entire package that is used in CheckFeerate.

Instead of PCKG_POLICY (which is supposed to be for package-wide
errors), use PCKG_TX. Also, to be more helpful to the caller, specify
which transactions were included in the evaluation and what the
feerate was.