WIP: sender-initiated 1P1C package relay
This defines a new "package inv" which
is defined as a pair of hashes that
are logically linked in a 1P1C relationship.

The receiving node can request the entire
package via getpkgtxns message, which is then
responded to in full by a pkgtxns message
or a notfound response.

This is meant to trigger IFF the 1P1C is a
chunk together, but as-is is triggered
when the parent is below the peer's feefilter
but the child is above it.

This potentially wastes 32 additional
bytes of relay traffice per package
per peer, and is likely incompatible
with erlay.

Possible extensions: only trigger this
logic if new package is going to be in
next block (to avoid orphanage churn),
or only trigger on TRUC transaction
packages.