fuzz: Add coverage for non-circular dep cluster Trim()'ing
When trimming, circular dependencies can be introduced via
the implicit dependencies generated (generated from the
existing linearization ordering in the cluster) as well as
the to-be-added dependencies.

If actual depenencies have not been eagerly computed, then
the implied dependency graph can contradict the to-be-added
dependencies, generating circular dependencies during a
Trim(). Therefore, one can have graphs that are completely
removed due to this cycle being introduced.

This commit adds a test case where the graph is
intentionally transitioning from non-oversized to
over-sized. Right before intentionally oversizing, the
to-be-added dependencies are applied by triggering graph
introspection, then all clusters are attached together
unidirectionally, to create a single large cluster that is
then trimmed.

The resulting trim is checked against a lower bound that
doesn't require implementing the Trim functionality again in
the simulator context.