save fuzz stuff with failing StichClusters
This was an attempt to avoid implicit cycles
using the text given in the PR as guidance. I take clusters
and then stitch them together in a single direction.

I couldn't convince myself the property held so I ended up
writing a test to try and convince myself.

One example failing input for StitchClusters2 where the
entire cluster is wiped out via cycle:

EhBGLQYGhDOENwYAr6+vr682AAYAAHoAAAAAAEr/EIQGBi0GBoQ0BjYQLUYGBoQzAEAzhAYGhDcG
ADIABgAAMwAAMQCEBgY=

This test was meant to make sure that everything was
"undersized" directly before, and use it like we'd imagine
it in the wild with reorgs.