script: introduce verify flags for OP_TEMPLATE
Script validation of OP_TEMPLATE is enabled past its activation height, which means it is currently
never enabled on mainnet/testnet and always is on regtest.

Tapscript spends making use of OP_TEMPLATE do not get relayed nor included in blocks until the soft
fork is active. Note how the standardness script verify flags are set such as even post activation,
an OP_TEMPLATEHASH related error would not be reported as a consensus error (and lead to peer
disconnection). This can be changed in a future release, once the soft fork activation isn't recent
anymore.

Co-Authored-by: Greg Sanders <gsanders87@gmail.com>