script: introduce verify flags for OP_TEMPLATEHASH
Script validation of OP_TEMPLATEHASH is enabled past
its activation height, which means it is currently never
enabled on mainnet/testnet and always is on regtest.

Tapscript spends making use of OP_TEMPLATEHASH do not get
relayed nor included in blocks until the soft fork is
active. We no longer disconnect or ban for loose mempool
transaction failures, so we can add the new flag directly to
mandatory flag set, which makes disconnect and bans possible
in blocks only post-activation.

Co-Authored-by: Greg Sanders <gsanders87@gmail.com>